+++ src/rateEngine/billingDeterminants/MonthlyDemand.ts
@@ -2,6 +2,9 @@ import BillingDeterminants from './_BillingDeterminants';
 import LoadProfile from '../LoadProfile';
 import { RateElementClassification, BillingDeterminantsUnits, RateElementTypeEnum } from '../constants';
 
+/**
+ * @deprecated Use the unified `Demand` class instead with demandPeriod: "monthly".
+ */
 class MonthlyDemand extends BillingDeterminants {
   private _loadProfile: LoadProfile;
 
@@ -11,12 +14,14 @@ class MonthlyDemand extends BillingDeterminants {
 
   constructor(loadProfile: LoadProfile) {
     super();
-
-    this._loadProfile = loadProfile;
+    this._loadProfile = loadProfile
+    throw new Error(
+      'MonthlyDemand has been deprecated. Please use RateElementTypeEnum.Demand with demandPeriod: "monthly".',
+    );
   }
 
   calculate(): Array<number> {
-    return this._loadProfile.byMonth().map(monthOfLoads => Math.max(...monthOfLoads));
+    return this._loadProfile.byMonth().map((monthOfLoads) => Math.max(...monthOfLoads));
   }
 }